chore(libs): update libs (jasmine, jstd, jasmine-jstd adapter)
