+++ lib/jasmine/version.txt
@@ -1,0 +1,1 @@
+1.1.0
