+++ lib/jstestdriver/version.txt
@@ -1,1 +1,1 @@
-1.3.2
+1.3.3c
